+++ apps/dashboard/src/app/account/invite-collaborators-modal/index.ts
@@ -1,0 +1,1 @@
+export * from './invite-collaborators-modal.component';