+++ libs/common/src/lib/kb-add/index.ts
@@ -1,2 +1,2 @@
 export * from './kb-add-modal.component';
-export * from './kb-creation-form/kb-creation-form.component';
+export * from './kb-creation-form';